+++ CHANGELOG.md
@@ -4,17 +4,40 @@
 
 ### Added
 
+- Issue templates for Github.
+- Add `transparentColor` and `useTransparentColor` to allow the layer to set a color to be "transparent" (or use the automatically calculated one when colormaps are set).
+
+### Changed
+
+## 0.8.2
+
+### Added
+
+### Changed
+
+- Fix opacity with photometric interpretations.
+- Only show background image when opacity is 1 (and viewport id prop matches that of the current viewport).
+- Fix `minZoom` calculation bug in `MultiscaleImageLayer`.
+
+## 0.8.1
+
+### Added
+
+- Property `onViewStateChange` of all viewers accepts a callback for deck.gl view state changes.
+- Add a click handler to the overview that centers the detail view on the click position. The handler can be turned off by setting the `PictureInPictureViewer` property `clickCenter` to `false`.
+- Support interleaved RGB OME-TIFF files.
+
 ### Changed
 
 - Update `getDefaultInitialViewState` to return floating point zoom that fills the screen by default.
 - Fix documentation of instances where spreading is used in the arguments for a function.
+- Fix link to bioformats in README
 
 ## 0.8.0
 
 ### Added
 
-- Export a `getDefaultInitialViewState` function for getting a deafult inital view state given a loader and desired size of view.
-- Support interleaved RGB OME-TIFF files.
+- Export a `getDefaultInitialViewState` function for getting a default initial view state given a loader and desired size of view.
 - Added a README file for Avivator at `avivator/README.md`.
 
 ### Changed